requirements.txt
requests
pyTelegramBotAPI
PyYAML
pycryptodome
il
